requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
